Resolve some issues in error defitions, and implement for prc
These errors use boxed errors for sub errors, which can behave a
little strangely. For now the code assumes that refernces to errors
are static, and it seems to be compiling. If this is an issue in the
future, it may be easier to remove these than add expicit lifetimes.